pyproject.toml
[project]
name = "wxauto4"
version = "40.1.2"
description = "wxauto for 4.0 client"
authors = [
    {name = "Cluic", email = "tikic@qq.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.14"
dependencies = [
    "tenacity",
    "pywin32",
    "pyperclip",
    "pillow",
    "psutil",
    "colorama",
    "comtypes",
    "websocket-client"
]

[project.scripts]
wxauto4 = "wxauto4.__main__:main"
wxauto4-bridge = "wxauto4.bridge.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
